rename isfc to isc
